+++ src/RequestInformation.cs
@@ -124,6 +124,10 @@ namespace Microsoft.Kiota.Abstractions
             bool boolean => boolean.ToString().ToLower(),// pass in a lowercase string as the final url will be uppercase due to the way ToString() works for booleans
             DateTimeOffset dateTimeOffset => dateTimeOffset.ToString("o"),// Default to ISO 8601 for datetimeoffsets in the url.
             DateTime dateTime => dateTime.ToString("o"),// Default to ISO 8601 for datetimes in the url.
+#if NET6_0_OR_GREATER
+            DateOnly dateOnly => dateOnly.ToString("yyyy-MM-dd"),// Default to ISO 8601 for dateonlys in the url.
+            TimeOnly timeOnly => timeOnly.ToString(@"HH\:mm\:ss"),// Default to ISO 8601 for timeonlys in the url.
+#endif
             Guid guid => guid.ToString("D"),// Default of 32 digits separated by hyphens
             Date date => date.ToString(), //Default to string format of the custom date object
             Time time => time.ToString(), //Default to string format of the custom time object